+++ package.js
@@ -1,15 +1,15 @@
 /* global Package */
 Package.describe({
   name: 'freedombase:legal-management',
-  version: '1.5.0',
+  version: '1.6.0',
   summary: 'Manage your legal documents and user consent.',
   git: 'https://github.com/freedombase/meteor-legal-management',
   documentation: 'README.md'
 })
 
 Package.onUse(function (api) {
-  api.versionsFrom('2.0')
-  api.use(['meteor', 'ecmascript', 'check', 'mongo', 'typescript'])
+  api.versionsFrom('2.3')
+  api.use(['meteor', 'ecmascript', 'check', 'mongo', 'typescript', 'callback-hook'])
   api.use(['aldeed:collection2@3.5.0', 'aldeed:schema-index@3.0.0', 'aldeed:schema-deny@3.0.0'])
   // 'socialize:base-model@1.1.2'